Create new Source Plugin that provides a fieldable entity. Create new profile migration for OSU Faculty profile. Update OSU Person profile to use new source plugin. Remove osu person source plugin no longer needed.
